Move depthInput to components
