+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/lithologyTable.tsx
@@ -7,7 +7,6 @@ import { AddRowButton } from "../../../../../../components/buttons/buttons.tsx";
 import { FaciesDescription } from "../../faciesDescription.ts";
 import { LithologicalDescription } from "../../lithologicalDescription.ts";
 import { Lithology } from "../../lithology.ts";
-import { DepthInput } from "../../lithology/depthInput.tsx";
 import { FaciesDescriptionLabels } from "../../lithology/faciesDescriptionLabels.tsx";
 import { FaciesDescriptionModal } from "../../lithology/form/faciesDescriptionModal.tsx";
 import { LithologicalDescriptionModal } from "../../lithology/form/lithologicalDescriptionModal.tsx";
@@ -15,6 +14,7 @@ import { LithologyModal } from "../../lithology/form/lithologyModal.tsx";
 import { LithologyLabels } from "../../lithology/lithologyLabels.tsx";
 import { DepthColumnCell } from "../depthColumnCell.tsx";
 import { DepthDeleteButton } from "../depthDeleteButton.tsx";
+import { DepthInput } from "../depthInput.tsx";
 import { DepthInsertButton } from "../depthInsertButton.tsx";
 import { LayerAddButton } from "../layerAddButton.tsx";
 import { StratigraphyTableActionCell } from "../stratigraphyTableActionCell.tsx";
